+++ src/main/java/org/ojalgo/matrix/BasicMatrix.java
@@ -23,7 +23,6 @@ package org.ojalgo.matrix;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Supplier;
 
 import org.ojalgo.ProgrammingError;
 import org.ojalgo.algebra.NormedVectorSpace;
@@ -38,10 +37,8 @@ import org.ojalgo.matrix.decomposition.LU;
 import org.ojalgo.matrix.decomposition.MatrixDecomposition;
 import org.ojalgo.matrix.decomposition.QR;
 import org.ojalgo.matrix.decomposition.SingularValue;
-import org.ojalgo.matrix.store.ElementsSupplier;
 import org.ojalgo.matrix.store.MatrixStore;
 import org.ojalgo.matrix.store.PhysicalStore;
-import org.ojalgo.matrix.store.PhysicalStore.Factory;
 import org.ojalgo.matrix.store.TransformableRegion;
 import org.ojalgo.matrix.task.DeterminantTask;
 import org.ojalgo.matrix.task.InverterTask;
@@ -49,7 +46,6 @@ import org.ojalgo.matrix.task.SolverTask;
 import org.ojalgo.scalar.Scalar;
 import org.ojalgo.structure.Access1D;
 import org.ojalgo.structure.Access2D;
-import org.ojalgo.structure.Mutate2D;
 import org.ojalgo.structure.Structure2D;
 import org.ojalgo.type.NumberDefinition;
 import org.ojalgo.type.context.NumberContext;
@@ -110,7 +106,6 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     }
 
     private transient MatrixDecomposition<N> myDecomposition = null;
-    private final PhysicalStore.Factory<N, ?> myFactory;
     private transient int myHashCode = 0;
     private transient Boolean myHermitian = null;
     private transient Boolean mySPD = null;
@@ -122,42 +117,19 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
         super();
 
         myStore = store;
-        myFactory = store.physical();
     }
 
     public M add(final double scalarAddend) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        N right = physical.scalar().cast(scalarAddend);
-
-        retVal.modifyAll(physical.function().add().second(right));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.add(scalarAddend));
     }
 
     public M add(final M addend) {
-
         ProgrammingError.throwIfNotEqualDimensions(myStore, addend);
-
-        PhysicalStore<N> retVal = myStore.physical().copy(addend);
-
-        retVal.modifyMatching(myStore, myStore.physical().function().add());
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.add(addend.getStore()));
     }
 
     public M add(final N scalarAddend) {
-
-        PhysicalStore.Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        retVal.modifyAll(physical.function().add().second(scalarAddend));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.add(scalarAddend));
     }
 
     public N aggregateColumn(final long row, final long col, final Aggregator aggregator) {
@@ -181,9 +153,13 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     }
 
     /**
-     * @return A fully mutable matrix builder with the elements initially set to a copy of this matrix.
+     * The returned instance can be have its elements mutated in various ways, while the size/shape is fixed.
+     *
+     * @return A fully mutable matrix builder with the elements initially set to a copy of this matrix –
+     *         always creates a full dense copy.
+     * @see #logical()
      */
-    public abstract <R extends Mutate2D.ModifiableReceiver<N> & Supplier<M>> R copy();
+    public abstract Mutator2D<N, M, PhysicalStore<N>> copy();
 
     public long count() {
         return myStore.count();
@@ -198,35 +174,19 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     }
 
     public M divide(final double scalarDivisor) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        N right = physical.scalar().cast(scalarDivisor);
-
-        retVal.modifyAll(physical.function().divide().second(right));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.divide(scalarDivisor));
     }
 
     public M divide(final N scalarDivisor) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        retVal.modifyAll(physical.function().divide().second(scalarDivisor));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.divide(scalarDivisor));
     }
 
     public double doubleValue(final long index) {
         return myStore.doubleValue(index);
     }
 
-    public double doubleValue(final long i, final long j) {
-        return myStore.doubleValue(i, j);
+    public double doubleValue(final long row, final long col) {
+        return myStore.doubleValue(row, col);
     }
 
     public M enforce(final NumberContext context) {
@@ -276,8 +236,8 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
         return myStore.get(index);
     }
 
-    public N get(final long aRow, final long aColumn) {
-        return myStore.get(aRow, aColumn);
+    public N get(final long row, final long col) {
+        return myStore.get(row, col);
     }
 
     /**
@@ -398,46 +358,32 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
         return mySymmetric.booleanValue();
     }
 
+    /**
+     * Compared to {@link #copy()} this does not create a copy – not initially anyway. The returned instance
+     * is a starting point for logically composing a new matrix.
+     *
+     * @return A logical builder that tries to avoid unnecessary copying.
+     * @see #copy()
+     */
     public abstract Pipeline2D<N, M, ?> logical();
 
     public M multiply(final double scalarMultiplicand) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        N right = physical.scalar().cast(scalarMultiplicand);
-
-        retVal.modifyAll(physical.function().multiply().second(right));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.multiply(scalarMultiplicand));
     }
 
     public M multiply(final M multiplicand) {
 
         ProgrammingError.throwIfMultiplicationNotPossible(myStore, multiplicand);
 
-        return this.newInstance(myStore.multiply(this.cast(multiplicand).collect(myFactory)));
+        return this.newInstance(myStore.multiply(multiplicand.getStore()));
     }
 
     public M multiply(final N scalarMultiplicand) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        retVal.modifyAll(physical.function().multiply().second(scalarMultiplicand));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.multiply(scalarMultiplicand));
     }
 
     public M negate() {
-
-        PhysicalStore<N> retVal = myStore.copy();
-
-        retVal.modifyAll(myStore.physical().function().negate());
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.negate());
     }
 
     /**
@@ -456,11 +402,11 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     }
 
     public M reduceColumns(final Aggregator aggregator) {
-        return this.newInstance(myStore.reduceColumns(aggregator).collect(myFactory));
+        return this.newInstance(myStore.reduceColumns(aggregator).collect(myStore.physical()));
     }
 
     public M reduceRows(final Aggregator aggregator) {
-        return this.newInstance(myStore.reduceRows(aggregator).collect(myFactory));
+        return this.newInstance(myStore.reduceRows(aggregator).collect(myStore.physical()));
     }
 
     public M signum() {
@@ -488,38 +434,16 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     }
 
     public M subtract(final double scalarSubtrahend) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        N right = physical.scalar().cast(scalarSubtrahend);
-
-        retVal.modifyAll(physical.function().subtract().second(right));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.subtract(scalarSubtrahend));
     }
 
     public M subtract(final M subtrahend) {
-
         ProgrammingError.throwIfNotEqualDimensions(myStore, subtrahend);
-
-        PhysicalStore<N> retVal = myStore.physical().copy(subtrahend);
-
-        retVal.modifyMatching(myStore, myStore.physical().function().subtract());
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.subtract(subtrahend.getStore()));
     }
 
     public M subtract(final N scalarSubtrahend) {
-
-        Factory<N, ?> physical = myStore.physical();
-
-        PhysicalStore<N> retVal = physical.copy(myStore);
-
-        retVal.modifyAll(physical.function().subtract().second(scalarSubtrahend));
-
-        return this.newInstance(retVal);
+        return this.newInstance(myStore.subtract(scalarSubtrahend));
     }
 
     public void supplyTo(final TransformableRegion<N> receiver) {
@@ -642,8 +566,6 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
 
         return task.toSolutionProvider(myStore, rhs);
     }
-
-    abstract ElementsSupplier<N> cast(Access1D<?> matrix);
 
     MatrixStore<N> getStore() {
         return myStore;