+++ src/test/java/org/ojalgo/matrix/P20030528Case.java
@@ -38,7 +38,7 @@ import org.ojalgo.type.context.NumberContext;
  */
 public class P20030528Case extends BasicMatrixTest {
 
-    private static final NumberContext DEFINITION = new NumberContext(7, 1);
+    private static final NumberContext DEFINITION = NumberContext.of(7, 1);
 
     public static RationalMatrix getProblematic() {
         final RationalMatrix tmpMtrx = RationalMatrix.FACTORY