+++ src/test/java/org/ojalgo/matrix/P20050125Case.java
@@ -39,7 +39,7 @@ import org.ojalgo.type.context.NumberContext;
  */
 public class P20050125Case extends BasicMatrixTest {
 
-    private static final NumberContext DEFINITION = new NumberContext(7, 9);
+    private static final NumberContext DEFINITION = NumberContext.of(7, 9);
 
     public static RationalMatrix getProblematic() {
         int DIM = 3;