+++ src/test/java/org/ojalgo/matrix/SimpleEquationCase.java
@@ -29,7 +29,7 @@ import org.ojalgo.type.context.NumberContext;
  */
 public class SimpleEquationCase extends BasicMatrixTest {
 
-    private static final NumberContext DEFINITION = new NumberContext(7, 1);
+    private static final NumberContext DEFINITION = NumberContext.of(7, 1);
 
     public static RationalMatrix getBody() {
         final RationalMatrix tmpMtrx = RationalMatrix.FACTORY.rows(new double[][] { { 2.0, 1.0, 1.0 }, { 4.0, -6.0, 0.0 }, { -2.0, 7.0, 2.0 } });