+++ src/test/java/org/ojalgo/matrix/SimpleLeastSquaresCase.java
@@ -34,7 +34,7 @@ import org.ojalgo.type.context.NumberContext;
  */
 public class SimpleLeastSquaresCase extends BasicMatrixTest {
 
-    private static final NumberContext DEFINITION = new NumberContext(7, 4);
+    private static final NumberContext DEFINITION = NumberContext.of(7, 4);
 
     public static RationalMatrix getBody() {
         final RationalMatrix tmpMtrx = RationalMatrix.FACTORY